pyproject.toml
[project]
name = "gym-guide-chatbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ali Hassan",email = "alihassan060104@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<3.9.7 || >3.9.7,<3.14"
dependencies = [
    "langchain-openai (>=0.3.28,<0.4.0)",
    "langchain-community (>=0.3.27,<0.4.0)",
    "langchain-pinecone (>=0.2.11,<0.3.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "pypdf (>=5.8.0,<6.0.0)",
    "sentence-transformers (>=5.0.0,<6.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-huggingface (>=0.3.1,<0.4.0)",
    "streamlit (>=1.47.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "gym_guide_chatbot", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
